1. Update/upgrade function works now In the original script awk was used like this: ```bash awk -F"/" "{print $1}" ``` In bash, when you use double quotes, variable substitution is allowed. That is, if there is a $1 or any variable inside the double quotes, bash will try to substitute it with the variable's value. In your case, $1 was not a bash variable but intended for awk. So bash was probably substituting $1 with nothing before awk had a chance to process it.
Changing the double quotes to single quotes:
```bash
awk -F/ '{print $1}'
```
prevents this from happening, because single quotes in bash will preserve the literal value of all characters within the quotes.
This means the $1 is passed correctly to awk, allowing it to operate as intended.

The backslashes (\') in the script are used to escape the single quotes so that they are correctly interpreted within the outer
set of single quotes defining the function.

2. Disable .NET telemetry

3. Stupid ubuntu runs like normal when lid is closed. No more battery discharges... Right?